more typescript changes

+++ src/app/api/categories/updatecategory/route.ts
@@ -1,47 +1,42 @@
-
 import { NextRequest, NextResponse } from "next/server";
 import { PrismaClient } from '@prisma/client';
-import { getDataFromToken } from "~/helpers/getDataFromToken";
+import { getDataFromToken} from "~/helpers/getDataFromToken";
 
 const prisma = new PrismaClient();
 
-export async function POST(request: NextRequest, response:NextResponse) {
+export async function POST(request: NextRequest, response: NextResponse) {
   try {
-    
-    
-    const userId : any = await getDataFromToken(request);
+    const userId = await getDataFromToken(request); 
 
     const reqBody = await request.json();
-    const { id  } : any  = reqBody;
+    const { id }: { id: number } = reqBody; 
 
-    const category = await prisma.categoriesOnUsers.findUnique({
-      where: { userId_categoryId: { userId, categoryId: id } }, 
-    });
-
-    let Row;
-
-    if(category === null){
-
-     Row = await prisma.categoriesOnUsers.create({
-      data: {
-        user: { connect: { id: userId } }, // Use connect to link to existing user
-        category: { connect: { id: id } }, // Use connect to link to existing category
-        assignedBy: 'YOUR_ASSIGNED_BY_VALUE', // Assign a value for assignedBy
-      },
-    });
-  } else{
-     Row = await prisma.categoriesOnUsers.delete({
+    const existingCategory = await prisma.categoriesOnUsers.findUnique({
       where: { userId_categoryId: { userId, categoryId: id } },
     });
-  }
-      
-    
+
+    let row: { categoryId: number, userId: number};
+
+    if (existingCategory === null) {
+      row = await prisma.categoriesOnUsers.create({
+        data: {
+          user: { connect: { id: userId } }, // Use connect to link existing user
+          category: { connect: { id } }, // Use connect to link existing category
+          assignedBy: 'YOUR_ASSIGNED_BY_VALUE', // Replace with appropriate value
+        },
+      });
+    } else {
+      row = await prisma.categoriesOnUsers.delete({
+        where: { userId_categoryId: { userId, categoryId: id } },
+      });
+    }
 
     return NextResponse.json({
-      message: "Categories found",
-      Row
+      message: existingCategory ? "Category unassigned" : "Category assigned",
+      Row: row,
     });
-  } catch (error: any) {
-    return NextResponse.json({ error: error.message }, { status: 400 });
-  } 
+  } catch (error) {
+    console.error("Error handling category association:", error.message);
+    return NextResponse.json({ error: "Internal server error" }, { status: 500 });
+  }
 }
